# openapi.yaml
openapi: 3.0.3
info:
  title: SHOBOT AMR API
  description: >
    REST API for managing SHOBOT AMRs including users, maps,
    waypoints, missions, zones, graphs, robot control, and analytics.
  version: 1.0.0
  contact:
    name: IIT Madras
  license:
    name: Proprietary
servers:
  - url: http://{host}:{port}/api/v1
    variables:
      host:
        default: localhost
      port:
        default: "8080"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Map:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        isActive:
          type: boolean
...
    Graph:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        nodes:
          type: array
          items:
            type: object
        edges:
          type: array
          items:
            type: object
security:
  - BearerAuth: []
paths:
...
  /maps:
    get:
      tags: [Maps]
      summary: Get all maps
      responses:
        "200":
          description: Maps list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Map'

    post:
      tags: [Maps]
      summary: Create map
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Map'
      responses:
        "201":
          description: Map created
  /maps/load/{mapId}:
    post:
      tags: [Maps]
      summary: Load map into robot
      parameters:
        - name: mapId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Map loaded

  # ---------------- WAYPOINTS ----------------
...
  /graphs:
    get:
      tags: [Graphs]
      summary: Get all graphs
      responses:
        "200":
          description: Graphs list

    post:
      tags: [Graphs]
      summary: Create graph
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Graph'
      responses:
        "201":
          description: Graph created
  /graphs/{graphId}:
    delete:
      tags: [Graphs]
      summary: Delete graph
      parameters:
        - name: graphId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Graph deleted

  # ---------------- ROBOT ----------------
  /robot/quickgoal/{waypointId}:
    post:
      tags: [Robot]
      summary: Send robot to waypoint
      parameters:
        - name: waypointId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Robot moving

  # ---------------- LOGS ----------------
  /logs:
    get:
      tags: [Logs]
      summary: Get logs
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Logs retrieved

  # ---------------- ANALYTICS ----------------
...